requirements.txt
Django==3.2.16
psycopg2==2.8.6
djangorestframework==3.12.2
django-cors-headers==3.13.0
drf-spectacular==0.25.1
python-dotenv==0.21.1